Fix: when window is resized, some buttons lose user feedback and stop working as expected (not updating UI). Solved changing the way layout children element boundaries are applied in application drawing cycle.
